[Wb] take cfg as argument instead of loading it a second time
