+++ scripts/mlp/wholebody/croccodyl.py
@@ -1,5 +1,4 @@
-import mlp.config as cfg
 
 
-def generateWholeBodyMotion(cs, fullBody=None, viewer=None):
+def generateWholeBodyMotion(cs, cfg, fullBody=None, viewer=None):
     raise NotImplemented("TODO")
